tests/krate/publish: Extract `basics` module
